# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Telegram Channel Manager API
  description: |
    # 🚀 Telegram Channel Manager v2.2
    
    سیستم حرفه‌ای مدیریت کانال‌های تلگرام با قابلیت‌های پیشرفته
    
    ## 🔑 احراز هویت
    تمام endpoint ها نیاز به کلید API دارند که باید در header ارسال شود:
    ```
    x-api-key: your_api_key_here
    ```
    
    ### دریافت کلید API
    1. با Master Key به `/api/keys/generate` درخواست POST بدید
    2. کلید API و Recovery Token رو ذخیره کنید
    3. در صورت گم شدن کلید، با Recovery Token می‌تونید بازیابی کنید
    
    ## ⚡ محدودیت‌های Rate Limit
    | عملیات | محدودیت | پنجره زمانی |
    |--------|----------|-------------|
    | Join Channel | 5 | دقیقه |
    | Leave Channel | 10 | دقیقه |
    | Channel Info | 30 | دقیقه |
    | Channel List | 10 | دقیقه |
    | Cleanup | 2 | ساعت |
    | Global | 100 | دقیقه |
    
    ## 📊 Session Management
    سشن‌ها از فایل `sessions.json` خوانده می‌شوند (مسیر قابل تنظیم در `SESSIONS_FILE_PATH`)
    
    ## 🔄 Response Codes
    - `200` - Success
    - `400` - Bad Request
    - `401` - Unauthorized (Invalid API Key)
    - `403` - Forbidden (No Permission)
    - `429` - Rate Limit Exceeded
    - `500` - Internal Server Error
    - `503` - Service Unavailable

  version: 2.2.0
  contact:
    name: Support Team
    email: support@telegram-manager.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development Server
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/keys/generate:
    post:
      tags: [Authentication]
      summary: Generate API Key
      description: Generate new API key with recovery token (Master key required)
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: "Production API Key"
                description:
                  type: string
                  example: "Key for production environment"
      responses:
        '200':
          description: API key created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  apiKey:
                    type: string
                    description: "Save this - shown only once!"
                  recoveryToken:
                    type: string
                    description: "Use this to recover lost API key"
                  id:
                    type: integer
                  name:
                    type: string
                  message:
                    type: string
  /api/keys/recover:
    post:
      tags: [Authentication]
      summary: Recover API Key
      description: Recover lost API key using recovery token (PUBLIC - no auth needed)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, recoveryToken]
              properties:
                name:
                  type: string
                  example: "Production API Key"
                recoveryToken:
                  type: string
                  example: "recovery123..."
      responses:
        '200':
          description: API key recovered
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  apiKey:
                    type: string
                  message:
                    type: string
...
  /api/session/status:
    get:
      tags: [Sessions]
      summary: Get Sessions Status
      description: Get detailed status of all Telegram sessions
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Sessions status information
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
...
  /api/channel/join:
    post:
      tags: [Channels]
      summary: Join Channel
      description: |
        Join a Telegram channel/group
        
        **Rate Limit:** 5 requests per minute
        
        **Supported formats:**
        - Username: `@channelname`
        - Public link: `https://t.me/channelname`
        - Private invite: `https://t.me/joinchat/ABCD123`
        - Private invite: `https://t.me/+ABCD123`
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel]
              properties:
                channel:
                  type: string
                  example: "@testchannel"
      responses:
        '200':
          description: Successfully joined channel
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  sessionUsed:
                    type: string
                  channelId:
                    type: string
                  channelTitle:
                    type: string
                  channelUsername:
                    type: string
                  sessionCapacity:
                    type: string
                  remainingSlots:
                    type: integer
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/channel/leave:
    post:
      tags: [Channels]
      summary: Leave Channel
      description: |
        Leave a Telegram channel
        
        **Rate Limit:** 10 requests per minute
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channelId]
              properties:
                channelId:
                  type: string
                  example: "-1001234567890"
                sessionName:
                  type: string
                  description: Optional - specific session to use
      responses:
        '200':
          description: Successfully left channel
  /api/channel/info:
    get:
      tags: [Channels]
      summary: Get Channel Information
      description: |
        Get detailed information about a channel
        
        **Rate Limit:** 30 requests per minute
      security:
        - ApiKeyAuth: []
      parameters:
        - name: channel
          in: query
          required: true
          schema:
            type: string
          example: "@testchannel"
      responses:
        '200':
          description: Channel information
  /api/channel/list:
    get:
      tags: [Channels]
      summary: List All Channels
      description: |
        Get list of all channels across all sessions
        
        **Rate Limit:** 10 requests per minute
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: List of channels
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      total:
                        type: integer
                      bySession:
                        type: object
...
  /api/channel/cleanup:
    post:
      tags: [Channels]
      summary: Cleanup Inactive Channels
      description: |
        Leave inactive or low-quality channels automatically
        
        **Rate Limit:** 2 requests per hour
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                days:
                  type: integer
                  default: 7
      responses:
        '200':
          description: Cleanup completed

  # ═══════════════════════════════════════
  # Monitoring & Metrics
  # ═══════════════════════════════════════
  /api/monitoring/status:
    get:
      tags: [Monitoring]
      summary: Get Monitoring Status
      description: Get current monitoring service status
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Monitoring status
  /api/monitoring/realtime:
    get:
      tags: [Monitoring]
      summary: Get Realtime Metrics
      description: Get realtime system metrics
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Realtime metrics
...
  /api/ratelimit/status:
    get:
      tags: [RateLimit]
      summary: Get Rate Limit Status
      description: Check current rate limit status for your API key
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Rate limit status
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      client:
                        type: string
                      limits:
                        type: object
                      currentTime:
                        type: string
                        format: date-time
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        API Key required for authentication
        
        Get your key:
        1. Use master key to call POST /api/keys/generate
        2. Save the API key and recovery token
        3. Use the API key in x-api-key header
  schemas:
...
    RateLimitError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        retryAfter:
          type: integer
        limit:
          type: integer
        remaining:
          type: integer
        resetTime:
          type: string
          format: date-time
        operation:
          type: string
...
  responses:
...
    RateLimitExceeded:
      description: Rate Limit Exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitError'
...